requirements.txt
blinker==1.8.2
Bootstrap-Flask==2.3.3
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email-validator==1.3.0
Flask==2.3.2
Flask-CKEditor==0.5.1
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
SQLAlchemy==2.0.25
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
Werkzeug==3.0.0
WTForms==3.0.1


